Fix deadlock in scheduler
On slow system, it was possible to create a deadlock in scheduler
resulting in all scheduled jobs to stop running (top_info, checks,
facts, ...).

This occur when scheduler was processing its job and a signal HUP
happen. When scheduler process its job, it hold a lock. When SIGHUP is
processed, it try to acquire this lock. Since both code were running in the
same thread (main thread), the second lock acquire hangs forever blocking
the main thread.

Now, signal handler only set flags which are proccessed later by a recuring
job.