Fix issue with unexisting command
In lots of subprocess call, code handle case were command returned
non-zero returncode, but not case were command did not existed at all.
For example "sudo" is likely to not exists on Windows.

Now for subprocess call that are allowed to fail, also catch OSError
which include error like command-not-found or bad-architecture (32 vs
64).