Support metrics from nginx and apache on https
No longer validate certificate while fetching metrics for nginx and
apache. Metrics are usually fetched using http, but in case of
redirection, the agent may ends fetching them over https.

For Apache, make sure the port "80" is not present (e.g. url is
not http://127.0.0.1:80), because it could generate a redirection to
https://127.0.0.1:80 (httpS on port 80).