Fix issue with discovery and ephemeral process
An ephemeral process (e.g. postgresql process created to serve one
connection) could be wrongly though as running outside any container.
This could result in a service being discovered on the host while in
reality is only running in a container.
Add a check to not considerate any processes that terminated during
the processes discovery.