Fix Docker discovery issue when Docker is upgraded
Due to the way the agent reconnect to Docker, during an upgrade, it
could happen that a discovery was run before the reconnection to Docker
which cause all service running in Docker to be seen as running outside
container.
To fix this issue, try to establish Docker connection before running the
discovery.