Fix metric points blocked after registration of the metric
Due to a race condition, it could be possible that few metric points
stay blocked for too long waiting for metric registration while actually
the metric is registered.
This happened due to wrong order between the update of the cache
containing newly registered metric and call to method that retry sending
blocked metric points.