Really avoid to see services on the host while running in Docker
It seems possible to not see a process using Docker API even if we query
(just) after its create time.
Now we try to use cgroup to guess the Docker container. If a Docker like
cgroup is used but the associated container is not found, ignore
recently (3 seconds) created process.