Fix missing container during Docker startup
If a discovery is run when Docker start, it may fail to get container
information but still see service processes. This happend because on
startup Docker seems to start containers before accepting requests (it
accept connection, but fail to serve them within our 10s timeout).

This situation could happen on a server that has lots of services to
start and not too fast.

To avoid such issue, the update of Docker information done during the
discovery will retry connection to Docker if the error is a timeout.