Support scrapping Prometheus exporter
Currently scrapped metrics are only exposed on Prometheus exporter (i.e.
we are a proxy to exporter).
We add a label "glouton_job" to distinguish metrics, how-ever it's still
needed to have same type and description for metric with same name (e.g.
all "go_threads" need to use the same type & description).