Handle MQTT reconnection inside Glouton
This was to avoid race-condition with auto-reconnect implemented by
paho.mqtt.golang.
I didn't find a way to call Disconnect() safely if paho was doing an
auto-reconnect (and Glouton can't known that an auto-reconnect is in
progress).

It should also fix a Glouton bug, where the gorouting in charge to
re-enable MQTT connection were also in charge of publising message over
MQTT and waited for the message to be send. So when the MQTT connection
was disabled, this gorouting blocked forever on message publishing.