Add watchdog for MQTT connection and healthcheck
If the MQTT connection didn't worked on last hour, force a restart of
the whole MQTT client. Do this at most once every 4 hours.
If this fail for more than 36 hours, kill Glouton itself.

The above process is run in Glouton healthcheck function. Added a
watchdog that kill Glouton if the healthcheck failed to run for 15
minutes (it normally run every minute).