synchronizer: test sync via a mock API
* monitors from the API are now checked aginst the configuration,
instead of iterating over metrics (that means we traded a complexity of
O(len(metrics)) with a ~constant-time complexity of marshalling/unmarshalling
yaml, not sure which is better, but I believe that linear iteration is
probably faster with a reasonable amount of metrics).
All that being said, let's not forget that this piece of code should
disappear when the configuration will be provided exclusively by the
API, which remain the lmid-term goal.
* The functions that operate on the configuration are now only defined
on the pointer type, to prevent potential issues when casting (as both
config.Configuration and *config.Configuration used to implement the
types.Config interface).