probes: generate a default blackbox module for remote probes
We no longer require the monitor to be defined both locally and on the
API. Now, if an agent is defined locally in the config, we will scrape
it and expose the prometheus metrics on the /metrics endpoint, but we
won't push these metrics over MQTT. If an agent is defined via the API,
it will be exposed on /metrics AND pushed over MQTT, as expected.
Given that the API does not provide us with a module, we generate module
configuration "on the fly", using the scheme of the monitor to determine
the appropriate configuration.