Fix partial points send
If a point value was NaN, it caused all following points in the same
iteration to be dropped.
So if the NaN is always in the first one, it may cause nearly all points
to be dropped.

NaN value don't usually happen (our collector shouldn't generate NaN),
but custom Prometheus exporter could easily generate NaN.