Relax timing comparision from strict to loose.
The rationale for this seemingly silly change is the lack of granularity
of windows timers, which broke tests (and could potentially break real
world cases too).

The most common issue was with code following this pattern:

func (o *Object) myGreatFunction(maxAge time.Duration) {
    if time.Since(o.lastUpdateTime) > maxAge {
        o.doSomething()

        o.lastUpdateTime = time.Now()
    }
}

While I do not know the timers used by golang's time.Now() on Windows,
I strongly believe that this is not using HPET or anything like that.
The end result if that if you call myGreatFunction in a short interval
(calling it in a loop for example), time.Since() will return 0, and we
won't execute the code. This means that even if we provide maxAge with
a value of 0 (which means we want to refresh instantly), we may not run
the code in the condition.

Also:
* Windows performance counters: disable unused collectors and drop most metrics
* disable the 'mem' telegraf input plugin on windows, as we use win_perf_counters
to derive more precise metrics.