windows: add support for enumrating processes on x86
One issue we may run into is that a glouton running in i386 mode won't
be able to enumerate the full path or the username of a 64 bit process.
The reverse is possible, so we should ensure that the installer never
distributes a 32 bit glouton on 64 bit windows, something that the
installer already takes care of.