Fix race-condition in APT pending upgrade metrics
Glouton Dpkg::Post-Invoke was run before the one of update-notifier,
which cause the read from Glouton and the write from update-notifier to
run concurrently.
So Glouton might read the old value or the new value depending on how
fast are both tasks.
Now Glouton wait until the file is updated or 10 seconds.