Make sure that error aren't silenced in PID <-> container association
Previously when an error occurre while trying to get the container for
one PID, it still filled some cache and returned the error for one PID.
Then later PID may reuse the cache (which is likely incomplet due to the
previous error) and then failed without any error.
Now on error, the cache is still filled (to avoid re-querying the
container-runtime) but the error is added to the cache, thus subsequent
query for another PID will see and use the error.