Fix minimal delay for synchronization loop
This is an issue when a new agent is registered, the minimal delay was
kept to 0 and Glouton consumed too much CPU.
This issue was introduced in commit done yesterday. But the state
yesterday was that on error the minimal delay was kept to 0.

Now the minimal delay is already increased to ~15 seconds after first
loop.