Fix wrong hash anti-collision
We had some code for handling (unlikely) collision in metric
labels hash, but this code had bugs when a metric is deleted on a slot
with collision.

The collision is too much unlikely (it's the hash used by Prometheus and
Prometheus handle more metrics than
Glouton), so ignore any possible collision.