Fix pending updates detection on Windows
We relied on the current user name being "NT AUTHORITY\LOCAL SERVICE", however this only works if the system language is English because the username is localized. In French the username was "SERVICE LOCAL". I removed this condition since Glouton is always run as the Local Service user on Windows.