Remove port used in additional instance from discovered instance
When multiple instance of the same service (e.g. multiple Redis or PHP-FPM) are running
and listenning on multiple ports, the discovery will detect all ports and associate it
with discovery instance (i.e. the one with empty string as instance).

If you do define in configuration an additional instance associated with one port, this port
will now be removed for ports of the discovered instance.