Better handle container error in process discovery
We used to fallback on not associating process with container when
container runtime had errors, especially if error always happened.
This caused wrong discovery when a server rebooted and had high load
during boot.
We now ignore processes for which container runtime had error for a
delay. This delay depend on the error:
* errors that indicate that container runtime isn't used are delayed
  for shortest time (10 seconds)
* errors that are not understood are delayed for average time (1 mintue)
* errors that indicate timeout / precence of container runtime are
  delayed for long time (5 minute).