Copy label item in the annotation is alone
This avoid issue if an input generate metric with just an item as label,
because Bleemeo connector assume that when a metric only has an item as label,
then it use the annotation item (and not the label).
The bleemeo connector is probably wrong here and the annotation item would probably
be dropped to just use the label item. But this commit make fewer change and help
having everywhere: metric with just an item label == metric with an item annotation.