Don't use smartctl exit status for the device health
This generate too much false positive (it generate device="0" or
device="#" on some system). It feels that exit code = 1 is more
smartctl don't support the device (or the device don't support
SMART) than device had an issue.

Also in case of error, the only information we had was "smartctl exit code
is 1" which isn't very helpful.