Sort facts to always return the same in case of duplicated facts
Glouton update its facts by first creating a new fact with the new value, then
deleting the old facts. This means that in case of a crash between this two operation
we could ends with a duplicate fact for the same value.
Since the duplicate state.json is based on fact values change, it's important to don't generate
false facts change.