Dynamic discovery handle listenning addresses conflict
When two service run in the same containers, it was possible that
dynamic discovery associated the same listenning address to two
different services. This could result service check / metric gathering
to be run on the wrong address (e.g. a container with nginx & php-fpm
and you do php-fpm metric gathring on port 80 where nginx run).