Fix possible nil-pointer access
If an event arrive for a container not known, we could use a nil-pointer.
This was very unlikely before previous commit, because only health event
used the event.Container value, and health event normally occur on
container that are running (i.e. we received a start event which fill
the container value before).