Fix possible delay in synchronization of metrics
After a discovery we should check if metric need to be registered.
But there was a race condition where the time of the discovery could be
read before the discovery occured, and then the date for last sync was
took. This leads to a discovery being ignored.
Now make sure the date of last sync is took early to ensure such issue
no longer occur. Now the "race condition" is the other way, we increase
chance of doing one extra synchronization check which isn't an issue.