Fix issue when metric took too long to be registered
In some corner case, a metric with the same labels (i.e. for glouton the
same metric, since we identify a metric by its labels) that change its
annotation could be missing in synchronization. It would then take up
to multiple hours to fix the issue.
Glouton now check for metric to register when a metric change its
annotation.

The real use-case to produce the issue was:
* Have a service check on a service inside a container
  (it should be a service check, because service check only send points
  every minutes, making the bug more likely).
* The container is deleted and replaced by another container with
  the same name (but another ID)
* The synchronization of metric would be triggered but skip the
  service check metric, because its annotation container ID would
  still be the old one which is no longer registered
* When metric with new annotation is received, since it labels only
  contains container name, it won't be a new metric and won't trigger
  metric synchronization.
  The change is to now trigger metric synchronization when annotation of
  metric change.