Fix vSphere VM memory usage metric
The mem.usage.average metric gives the percentage of the host memory used by the VM, but we want the percentage of the VM memory used. Thus, we gather the amount of memory used by the VM and combine it with the VM memory size to evaluate the % of used VM memory.