Fix metric filter on Prometheus scraper
The filter is made to works on labels after all transformations.
The possible transformation are:
* GathererModifier
* Relabel when done dynamically on every points
* Relabel done per-gatherer
* Renamer

The final filter applied on gatherer is now done after all those
transformation.

The pre-filter (HintMetricFilter) is skipped if GathererModifier or
dynamic relabel are used.

The per-gatherer relabel only result in labels being merged, this is
now applied during HintMetricFilter.

The renamer is only used on SNMP is should be converted into
SimpleRule, we shouldn't enable AcceptAllowedMetricsOnly for SNMP before renamer
is dropped.