Make unhealthy check more tolerant
The watchdog will wait for the 4th failure because panic (instead of 3rd failure).
This idea is that during extrem load / if server freeze for extended period,
we might ends doing two check at the same time when server is back of life:
* one because we could be currently doing the watchdog function
* one because the ticker might have send a tick that watchdog didn't yet processed

Then at this point, since ticker try to align tick, the next tick could arrive
rather quickly, in less than 1 minutes. Which cause the 3rd failure seen by watchdog
in a short time. By waiting a 4th failure seen by watchdog, we should give at least one
minute for Glouton to because healthy.

This assume that the server extrem load is ended, or else glouton might never get
healthy. But in that case - server stay way too slow - it seems right to panic
due to unhealthy situation.