Allow vSphere gatherers to wait for agent config.
When a vSphere gatherer is created, it doesn't know its devices yet. We need to wait for a gathering to complete to be aware of all the devices. Then we can deduce which agent config is the more suitable for the gatherer (the one that has the smallest metric resolution), so we can update the gatherer interval with the right period.