Remove the labels server and source from various inputs.
NATS metrics contained a labels "server" which was the origin of metrics.
uWSGI metrics contained a labels "source" which was the origin of the metrics.
OpenLDAP metrics contained two labels "server" and "port" which was the origin
of the metrics.

But since glouton already add origin information in item, instance or container_name
label this was redudant.