fix: Drop BleemeoItem annotation
This fix some metrics that were not consistent between BleemeoItem annotation and
label item. This was the case for io_utilization on Windows.

Originally the BleemeoItem annotation versus item labels was intended to support storing
the item into a labels named "device", "path", "volume"... depending on what was measured.
It was originally needed because Bleemeo didn't supported arbitrary labels.
Now that Bleemeo support it, this distinction only cause trouble and mistake.

While fixing it, two other issue where seen:
* PostgreSQL detailled database when running on a container didn't worked:
  it was expected to have the container name concatenated with database name, due to
  a bug only the container name was kept.
* HAproxy had the same issue with backend name in place of database name