refactor: change how discovery is updated/notified
Rather than having caller decide when to update discovery / max age of the
discovery, the discovery is now a running component that decide by itself
when we can update. It will notify any update to subscriber.
It also possible to get immediatelly the latest discovery result and to request
an update.

This will allow discovery system to have more flexibility on when another discovery
will be run and ensure subscriber (service check & metric configuration) get notified
immediately of an updated discovery.