fix: log component leak when service are watched and discovery happen
When the following condition are present:
* Glouton run in a container (need symlink resolution)
* A *service* watch a log file which is a symlink (easiest case, a custom service configuration file)
* A discovery occur (e.g. a container is started/stopped)

Then an additional filereceiver component would be started, because code would believe the logfile isn't
already watched. The issue was that we checked if logfile is already watched using path before symlink resolution,
but then added the logfile to watched files after symlink resolution.

Now always do the symlink resolution early, to be consitent in which file path is used.